+++ src/header.tsx
@@ -18,7 +18,9 @@ import type {NavigationBreadcrumbsProps} from './navigation-breadcrumbs';
 
 type OverridableTextProps = {
     color?: TextPresetProps['color'];
+    /** @deprecated use decoration prop */
     textDecoration?: TextPresetProps['textDecoration'];
+    decoration?: TextPresetProps['decoration'];
     truncate?: TextPresetProps['truncate'];
 };