+++ src/index.tsx
@@ -42,6 +42,7 @@ export {default as MasterDetailLayout} from './master-detail-layout';
 export {default as NavigationBreadcrumbs} from './navigation-breadcrumbs';
 export {default as Text, Text1, Text2, Text3, Text4, Text5, Text6, Text7, Text8, Text9, Text10} from './text';
 export {default as Tag} from './tag';
+export type {TagType} from './tag';
 export {default as SectionTitle} from './section-title';
 export {Placeholder, AvatarPlaceholder} from './placeholder';
 export {RowList, Row, BoxedRowList, BoxedRow} from './list';
@@ -70,12 +71,14 @@ export {
     NavigationBarLogo,
 } from './navigation-bar';
 export {default as Image} from './image';
+export {default as Video} from './video';
 export {Carousel, CenteredCarousel, PageBullets} from './carousel';
 
 // Forms
 export {default as Form} from './form';
 export {default as Select} from './select';
 export {default as TextField} from './text-field';
+export {TextFieldBase} from './text-field-base';
 export {default as SearchField} from './search-field';
 export {default as EmailField} from './email-field';
 export {default as PhoneNumberField} from './phone-number-field';
@@ -91,7 +94,7 @@ export {default as DecimalField} from './decimal-field';
 export {default as PasswordField} from './password-field';
 export {default as DoubleField} from './double-field';
 export {default as IbanField} from './iban-field';
-export {useForm} from './form-context';
+export {useForm, useFieldProps} from './form-context';
 
 // Icons
 // @deprecated use IconCloseRegular instead